// src/enums.ts
export enum EventStatus {
    Pre = 'PRE',
    Live = 'LIVE',
    Removed = 'REMOVED',
}

export enum CompetitorType {
    Home = 'HOME',
    Away = 'AWAY',
}

export enum ScoreType {
    Current = 'CURRENT',
    Period1 = 'PERIOD_1',
    Period2 = 'PERIOD_2',
    Period3 = 'PERIOD_3',
    Period4 = 'PERIOD_4',
}

export enum Sport {
    Football = 'FOOTBALL',
    Basketball = 'BASKETBALL',
    Tennis = 'TENNIS',
}